added hostname and port to hfs

+++ pcawg-import-client/src/main/java/org/icgc/dcc/pcawg/client/core/HdfsFileWriter.java
@@ -1,5 +1,6 @@
 package org.icgc.dcc.pcawg.client.core;
 
+import lombok.Getter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -12,7 +13,6 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.Writer;
-import java.util.Optional;
 
 @Slf4j
 public class HdfsFileWriter extends Writer {
@@ -20,35 +20,39 @@ public class HdfsFileWriter extends Writer {
   private static final String FS_PARAM_NAME = "fs.defaultFS";
   private static final String HDFS = "hdfs";
 
-  private static final Configuration DEFAULT_CONF = new Configuration();
-  {
-    DEFAULT_CONF.set(FS_PARAM_NAME, HDFS+":///");
-  }
-
-  private final FileSystem hdfs;
-
   private final Writer internalWriter;
 
-//  private final String hdfsAddress;
-//  private final String hdfsPort;
+  @Getter
+  private final boolean append;
 
-  private static Configuration createConfiguration(Optional<String> optionalCoreConfigFilename, Optional<String> optionalHdfsConfigFilename){
+  private static String createUrl(String fileSystemName, String hostname, String port){
+    return fileSystemName+"://"+hostname+":"+port;
+  }
+
+  private static Configuration createConfiguration(String hostname, String port){
     val conf = new Configuration();
-    log.info("CoreConfigFilename: {}", optionalCoreConfigFilename.isPresent() ? optionalCoreConfigFilename.get() : "null");
-    log.info("HdfsConfigFilename: {}", optionalHdfsConfigFilename.isPresent() ? optionalHdfsConfigFilename.get() : "null");
-    optionalCoreConfigFilename.ifPresent(conf::addResource);
-    optionalHdfsConfigFilename.ifPresent(conf::addResource);
+    val baseUrl = createUrl(HDFS, hostname, port);
+    conf.set(FS_PARAM_NAME, baseUrl);
     return conf;
   }
 
-
   @SneakyThrows
-  private static OutputStream createNewOutputStream(Path file, Configuration conf){
-    val fs = file.getFileSystem(conf);
-    log.info("FileName1: {}", file.toString());
+  private static FileSystem createFileSystem(String hostname, String port){
+    val conf = createConfiguration(hostname, port);
+    return FileSystem.get(conf);
+  }
+
+  //TODO: clean this up of all logs
+  @SneakyThrows
+  private static OutputStream createNewOutputStream(FileSystem fs, Path file, boolean append){
+    log.info("FileName: {}", file.toString());
     if( fs.exists(file)){
-      log.info("The file [{}] EXISTS! Deleting it ", file.toString() );
-      fs.delete(file, true);
+      if (!append){
+        log.info("The file [{}] EXISTS! Deleting it ", file.toString() );
+        fs.delete(file, true);
+      } else {
+        log.info("The file [{}] EXISTS! Appending to it ", file.toString() );
+      }
     } else {
       log.info("The file [{}] DNE ", file.toUri().toString() );
       val parent = file.getParent();
@@ -60,43 +64,14 @@ public class HdfsFileWriter extends Writer {
     return fs.create(file);
   }
 
-
-  @SneakyThrows
-  private static OutputStream createNewOutputStream(FileSystem fs, String filename){
-    val file = new Path(filename);
-    log.info("FileName1: {}", file.toString());
-    if( fs.exists(file)){
-      log.info("The file [{}] EXISTS! Deleting it ", file.toString() );
-      fs.delete(file, true);
-    } else {
-      log.info("The file [{}] DNE ", file.toUri().toString() );
-      val parent = file.getParent();
-      if (! fs.exists(parent)){
-        log.info("The parent dir [{}] doesnt exist, will create...  ", parent.toUri().toString() );
-        fs.mkdirs(parent);
-      }
-    }
-    return fs.create(file);
-  }
-
-  @SneakyThrows
-  public HdfsFileWriter(FileSystem hdfs, String outputFilename) throws IOException {
-    log.info("Connecting to: {}", hdfs.getConf().get(FS_PARAM_NAME));
-    log.info("Outputting to: {} ", outputFilename);
-    val os = createNewOutputStream(hdfs, outputFilename);
-    this.internalWriter = new BufferedWriter(new OutputStreamWriter(os));
-    this.hdfs = hdfs;
-  }
-
-  @SneakyThrows
-  public HdfsFileWriter(String outputFilename, Optional<String> optionalCoreConfigFilename, Optional<String> optionalHdfsConfigFilename) throws IOException {
+  public HdfsFileWriter(String hostname, String port, String outputFilename, boolean append) throws IOException {
     val file = new Path(outputFilename);
-    val conf = createConfiguration(optionalCoreConfigFilename,optionalHdfsConfigFilename);
-    this.hdfs = file.getFileSystem(conf);
+    val hdfs = createFileSystem(hostname,port);
     log.info("Connecting to: {}", hdfs.getConf().get(FS_PARAM_NAME));
     log.info("Outputting to: {} ", outputFilename);
-    val os = createNewOutputStream(file,conf  );
+    val os = createNewOutputStream(hdfs, file, append);
     this.internalWriter = new BufferedWriter(new OutputStreamWriter(os));
+    this.append = append;
   }
